src/backend/requirements.txt
fastapi
redis
sqlalchemy
alembic
uvicorn[standard]
howlongtobeatpy
python-dotenv
httpx